# Projeto2026/api_dados/swagger.yaml
openapi: 3.0.3
info:
  title: EW API de Dados
  version: 1.0.0
  description: API de dados para produtos, encomendas, cores e categorias.
servers:
  - url: http://localhost:7789
    description: Servidor local de desenvolvimento
...
paths:
  /produtos:
    get:
      tags: [Produtos]
      summary: Listar produtos
      description: |
        Lista produtos com filtros opcionais. Os filtros suportados incluem categoria, cor, tamanho,
        intervalo de preço, pesquisa por nome, ordenação e paginação.
      security: []
      parameters:
        - in: query
          name: categoria
          schema:
            type: string
          description: Identificador da categoria. O servidor normaliza o valor para formato com hífens.
        - in: query
          name: cor
          schema:
            type: string
          description: Identificador da cor em hexadecimal armazenado em `opcoes.cores._id`.
        - in: query
          name: tamanho
          schema:
            type: string
          description: Tamanho a procurar em `opcoes.tamanhos`.
        - in: query
          name: minPreco
          schema:
            type: number
          description: Preço mínimo para filtrar variantes.
        - in: query
          name: maxPreco
          schema:
            type: number
          description: Preço máximo para filtrar variantes.
        - in: query
          name: sort
          schema:
            type: string
            enum: [preco_asc, preco_desc, nome_asc, nome_desc]
          description: Ordenação por preço ou nome.
        - in: query
          name: search
          schema:
            type: string
          description: Pesquisa parcial no nome do produto.
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
          description: Página da paginação.
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
          description: Número de registos por página.
      responses:
        '200':
          description: Lista de produtos.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProdutosListResponse'
        '500':
          $ref: '#/components/responses/ServerError'
    post:
      tags: [Produtos]
      summary: Criar produto
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [produto]
              properties:
                produto:
                  type: string
                  description: JSON serializado com os dados do produto.
      responses:
        '201':
          description: Produto criado com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Produto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /produtos/{id}:
    get:
      tags: [Produtos]
      summary: Obter produto por ID
      security: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Produto encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Produto'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    put:
      tags: [Produtos]
      summary: Atualizar produto
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [produto]
              properties:
                produto:
                  type: string
                  description: JSON serializado com os dados atualizados do produto.
      responses:
        '200':
          description: Produto atualizado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Produto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    delete:
      tags: [Produtos]
      summary: Apagar produto
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Produto removido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /encomendas:
    get:
      tags: [Encomendas]
      summary: Listar encomendas
      description: |
        Requer autenticação. Clientes só conseguem ver as próprias encomendas; administradores podem consultar todas.
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: userId
          schema:
            type: string
          description: Filtra pelo utilizador. Para clientes, o servidor ignora este valor e usa o ID do token.
        - in: query
          name: estado
          schema:
            type: string
            enum: [pendente, concluida, cancelada]
        - in: query
          name: dataInicio
          schema:
            type: string
            format: date
          description: Data inicial do intervalo.
        - in: query
          name: dataFim
          schema:
            type: string
            format: date
          description: Data final do intervalo.
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
        - in: query
          name: sort
          schema:
            type: string
          description: Campo de ordenação passado diretamente para o Mongoose.
      responses:
        '200':
          description: Lista de encomendas.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EncomendasListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
    post:
      tags: [Encomendas]
      summary: Criar encomenda
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EncomendaInput'
      responses:
        '201':
          description: Encomenda criada com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Encomenda'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /categorias:
    get:
      tags: [Categorias]
      summary: Listar categorias
      security: []
      responses:
        '200':
          description: Lista de categorias.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Categoria'
        '500':
          $ref: '#/components/responses/ServerError'
    post:
      tags: [Categorias]
      summary: Criar categoria
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoriaInput'
      responses:
        '201':
          description: Categoria criada com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Categoria'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        O middleware aceita o token no header `Authorization: Bearer <token>`, cookie `token` ou query string `?token=`.
  parameters:
    IdPath:
      in: path
      name: id
      required: true
      schema:
        type: string
      description: Identificador MongoDB do recurso.
  responses:
    BadRequest:
      description: Pedido inválido.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Não autenticado.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AuthErrorResponse'
    Forbidden:
      description: Sem permissões para executar a operação.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AuthErrorResponse'
    NotFound:
      description: Recurso não encontrado.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ServerError:
      description: Erro interno do servidor.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: Produto not found
        erro:
          type: string
          example: Token inválido ou expirado.
    AuthErrorResponse:
      type: object
      properties:
        erro:
          type: string
      example:
        erro: Acesso negado. Requer permissão de administrador.
    MessageResponse:
      type: object
      properties:
        message:
          type: string
...
    ProdutoBase:
      type: object
      properties:
        _id:
          type: string
        nome:
          type: string
        categoriaId:
          type: string
        imagemBase:
          type: string
        opcoes:
          type: object
          properties:
            cores:
              type: array
              items:
                $ref: '#/components/schemas/ProdutoColorOption'
            tamanhos:
              type: array
              items:
                type: string
            tampas:
              type: array
              items:
                type: string
        variantes:
          type: array
          items:
            $ref: '#/components/schemas/ProdutoVariant'
...
    Produto:
      allOf:
        - $ref: '#/components/schemas/ProdutoBase'
    ProdutosListResponse:
      type: object
      properties:
        produtos:
          type: array
          items:
            $ref: '#/components/schemas/Produto'
        total:
          type: integer
...
    EncomendaBase:
      type: object
      properties:
        _id:
          type: string
        userId:
          type: string
        produtos:
          type: array
          items:
            $ref: '#/components/schemas/EncomendaItem'
        precoTotal:
          type: number
        data:
          type: string
          format: date-time
        estado:
          type: string
          enum: [pendente, concluida, cancelada]
    EncomendaInput:
      allOf:
        - $ref: '#/components/schemas/EncomendaBase'
      required:
        - userId
        - produtos
        - precoTotal
...
    Encomenda:
      allOf:
        - $ref: '#/components/schemas/EncomendaBase'
    EncomendasListResponse:
      type: object
      properties:
        encomendas:
          type: array
          items:
            $ref: '#/components/schemas/Encomenda'
        total:
          type: integer
...
    CategoriaInput:
      type: object
      properties:
        _id:
          type: string
        label:
          type: string
      required:
        - label
    Categoria:
      type: object
      properties:
        _id:
          type: string
        label:
          type: string
...